+++ pkg/dart/pub/testdata/broken.lock
