+++ pkg/dart/pub/testdata/empty.lock
